% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPCpretreatment.R
\name{add_traj_to_las}
\alias{add_traj_to_las}
\title{Add trajectory to lidar data}
\usage{
add_traj_to_las(las, traj)
}
\arguments{
\item{las}{LAS object}

\item{traj}{sf object with column gpstime and Point Z geometries}
}
\value{
LAS object with additional attributes Xtraj, Ytraj, Ztraj, Ttraj
}
\description{
Interpolate trajectory to las points and add coordinates as extrabytes
}
