% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_metrics.R
\name{class_metrics}
\alias{class_metrics}
\title{Class counts}
\usage{
class_metrics(classes = c(1:6, 9, 17, 18, 64, 66, 67))
}
\arguments{
\item{classes}{Vector of classes to count
Default class specs are from https://data.geopf.fr/annexes/ressources/documentation/DC_LiDAR_HD_1-0.pdf.}
}
\value{
Formula for lidR::xxx_metrics
}
\description{
Class counts
}
\examples{
\dontrun{
las_file <- system.file("extdata", "M30_FontBlanche.laz", package = "lidarforfuel")
las <- lidR::readLAS(las_file)
# as cloud metrics
class_count <- lidR::cloud_metrics(las, class_metrics())
# as pixel metrics, i.e. rasterized
class_count <- lidR::pixel_metrics(las, class_metrics(), res = 10)
}

}
