% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{datetime_to_gpstime}
\alias{datetime_to_gpstime}
\title{convert datetime to gps time}
\usage{
datetime_to_gpstime(datetime, gpstime_ref = "2011-09-14 01:46:40")
}
\arguments{
\item{datetime}{POSIXct vector of datetime}

\item{gpstime_ref}{character. reference datetime}
}
\value{
numeric. vector of gps time
}
\description{
convert datetime to gps time
}
