% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffuelmetrics.R
\name{ffuelmetrics}
\alias{ffuelmetrics}
\title{Fuel metrics LiDAR}
\usage{
ffuelmetrics(profile_table, threshold)
}
\arguments{
\item{profile_table}{A table (e.g., data.frame or data.table) with two columns: H (height of the strata) and BD (bulk density in kg/m³).}

\item{threshold}{Numeric or character. Default is 0.012. A critical bulk density threshold used to identify different strata limits such as midstorey height, canopy base, and canopy top. Can be either: Numeric: a bulk density value (in kg/m³) or Character: a percentage of the maximum CBD value (cf example).}
}
\value{
A named numeric vector containing the computed fuel metrics.
}
\description{
Computes fuel metrics from a bulk density profile. Produces similar outputs to the fCBDprofile_fuelmetrics function. This function is intended for use with bulk density profiles derived from field data, MLS, TLS, or model outputs.
}
\details{
In addition to fuel metrics, this function also estimates structural metrics such as total height, plant area index above 1 meter (PAI_tot), and vertical complexity index (VCI) based on the plant area density profile.
}
\examples{
\donttest{
path2profile <- system.file("extdata","BD_profile.txt", package="lidarforfuel")

# read a table
Profile <- data.table::fread(path2profile)

# Run the function with a relative threshold
Fuel_metrics <- ffuelmetrics(profile_table=Profile,threshold="10\%")
# Run the function with an absolute threshold
Fuel_metrics <- ffuelmetrics(profile_table=Profile,threshold=0.02)

}
}
