% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffuelmetrics2.R
\name{ffuelmetrics2}
\alias{ffuelmetrics2}
\alias{ffuelmetrics2Rast}
\title{Fuel metrics from PAD profiles or bulk density profiles}
\usage{
ffuelmetrics2(
  PADval = NULL,
  CBDval = NULL,
  FMAcan = NULL,
  FMAshrub = NULL,
  zval,
  dz = 0.5,
  LadderFuelBDthresholds = c(0.01, 0.02, 0.05),
  CanopyHeightThreshold = 1,
  CanopyBDTreshFrac = 0.1,
  CanopyTopBDThresholds = c(0.0012, 0.01),
  CBHThresholdRatio = 1/5,
  bdmax = 1,
  M = 2.5,
  asTibble = FALSE
)

ffuelmetrics2Rast(
  PAD_rast = NULL,
  CBD_rast = NULL,
  FMAcan_rast = NULL,
  FMAshrub_rast = NULL,
  zval,
  dz,
  LadderFuelBDthresholds = c(0.01, 0.02, 0.05),
  CanopyHeightThreshold = 1,
  CanopyBDTreshFrac = 0.1,
  CanopyTopBDThresholds = c(0.0012, 0.01),
  CBHThresholdRatio = 1/5,
  bdmax = 1,
  M = 2.5,
  asTibble = FALSE
)
}
\arguments{
\item{PADval}{numeric vector. PAD values in m2/m3. Set to NULL to input directly CBDval.}

\item{CBDval}{numeric vector. CBD values in kg/m3, only used if PADval is NULL.}

\item{FMAcan}{numeric. Fuel mass area in kg/m2 for canopy}

\item{FMAshrub}{numeric. Fuel mass area in kg/m2 for shrub}

\item{zval}{numeric vector. Middle height of the PAD layers in m, e.g. for PAD_1_10, zval should be 10.5.}

\item{dz}{numeric. Vertical resolution of the PADval or CBDval in m, e.g. for PAD_1_10, dz should be 1.}

\item{LadderFuelBDthresholds}{numeric vector. Minimal bulk density thresholds for Understorey and Ladder fuels (kg/m3)}

\item{CanopyHeightThreshold}{numeric. Default minimal value of height in m to look for CBH and CTH, typically 1 to 3 m}

\item{CanopyBDTreshFrac}{numeric. Fraction threshold to combine CBDmin and CBDmax to look for CBH (0.5 is average, 0.3 is closer to CBDmin)}

\item{CanopyTopBDThresholds}{numeric vector. Minimal bulkDensity to look for CTH (canopy height but in terms of fuel<Height)}

\item{CBHThresholdRatio}{numeric. Ratio of total height above which we look for CBH}

\item{bdmax}{numeric. Max value of bulk density kg/m3 for canopy (above 1 m only)}

\item{M}{numeric. M factor from Perrakis 2025}

\item{asTibble}{logical. If TRUE, returns a tibble.}
}
\description{
Fuel metrics from PAD profiles or bulk density profiles
}
