% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_gpstime}
\alias{filter_gpstime}
\alias{.filter_gpstime}
\title{test if points are in the season and around a date mode}
\usage{
filter_gpstime(
  gpstime,
  months = 1:12,
  deviation_days = Inf,
  gpstime_ref = "2011-09-14 01:46:40"
)

.filter_gpstime(
  months = 1:12,
  deviation_days = Inf,
  gpstime_ref = "2011-09-14 01:46:40"
)
}
\arguments{
\item{gpstime}{numeric. vector of gps time in seconds}

\item{months}{numeric vector of months defining the expected season}

\item{deviation_days}{numeric. the number of days defining the valid range around the date mode}

\item{gpstime_ref}{character. Default = "2011-09-14 01:46:40". The datetime corresponding to gpstime=0, in order to retrieve the real datetime of points.
It is expected to be in timezone UTC. Default is "2011-09-14 01:46:40" which is the standard GPS Time (1980-01-06 00:00:00)
plus 1e9 seconds, as defined in LAS 1.4 specifications.}
}
\value{
logical vector with TRUE for points within the season and date mode +- deviation_days.
}
\description{
These functions return TRUE when points are in the season and around the date mode,
the date mode being computed on points within the season.

\itemize{
\item \code{filter_gpstime} applies to \code{\link[lidR]{filter_poi}}
\item \code{.filter_gpstime} returns a formula for \code{\link[lidarforfuel]{pixel_filter}}
}
}
\examples{
\dontrun{
library(lidR)
filter_poi(las, filter_gpstime(gpstime, months = 5:9, deviation_days = 14))
# same as
filter_poi(las, filter_gpstime(months = 5:9)) |>
  filter_poi(filter_gpstime(deviation_days = 14))
# but different from
filter_poi(las, filter_gpstime(deviation_days = 14)) |>
  filter_poi(filter_gpstime(months = 5:9))
}
}
