% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPCpretreatment.R
\name{fPCpretreatment}
\alias{fPCpretreatment}
\title{Point cloud pre-treatment for using fCBDprofile_fuelmetrics in pixels}
\usage{
fPCpretreatment(
  chunk,
  classify = FALSE,
  Height_filter = 60,
  start_date = "2011-09-14 01:46:40",
  season_filter = 1:12,
  deviation_days = Inf,
  plot_hist_days = FALSE,
  traj = NULL
)
}
\arguments{
\item{chunk}{LAS object, LAScluster chunk or character. If character, the path to a las (laz) file is expected.
Can be apply to a catalog see lidR::catalog_apply # nolint: line_length_linter.}

\item{classify}{logical (default is FALSE). Make a ground classification. Only if the original point cloud is not classified}

\item{Height_filter}{numeric. Default = 80. Height limit to remove noise point}

\item{start_date}{date. The reference datetime to retrieve datetime from relative gpstime of the laz.
It is expected to be in timezone UTC. Default is "2011-09-14 01:46:40" which is the standard GPS Time (1980-01-06 00:00:00)
plus 1e9 seconds, as defined in LAS 1.4 specifications.}

\item{season_filter}{numeric. A vector of integer for months to keep (e.g: 5:10 keep retunrs between may and october)}

\item{deviation_days}{numeric. Maximum number of days tolerated between the acquisition in a given point cloud (a tile or plot). Deactivated by default}

\item{plot_hist_days}{logical. Should the histogram of dates of acquisition be displayed. Default =FALSE}

\item{traj}{sf object. Trajectory covering the LAS chunk. The sf object is expected to have a column gpstime and Point Z geometries.
If NULL, the function will try to compute it.}
}
\value{
LAS object. A normalized point cloud (.laz) with several new attributes needed to run fCBDprofile_fuelmetrics, see details.
}
\description{
Function for preprocessing las (laz) files for use in fCBDprofile_fuelmetrics. This can be used in the catalog_apply lidR function. The pretreatment consists of normalizing the point cloud and adding various attributes: Plane position for each point (easting, northing, elevation), LMA (leaf mass area) and wood density (WD) by intersecting the point cloud with an LMA and WD map or by providing LMA and WD values.
}
\details{
The attributes added to the laz are:
\itemize{
\item Zref : Z value before normalization, i.e. altitude of the point
\item Easting, Northing, Elevation, Time that are the X,Y,Z coordinates of the sensor
and the its GPStime for each point (obtained from lidR::track_sensor()).
}
}
\examples{
\donttest{
path2laz <- system.file("extdata", "M30_FontBlanche.laz", package = "lidarforfuel")
M30_FontBlanche_pretreated <- fPCpretreatment(path2laz)
# displaying the new attributes in the las
names(M30_FontBlanche_pretreated)
}
}
