% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_mass_area.R
\name{fuel_mass_area}
\alias{fuel_mass_area}
\title{Compute fuel mass area from leaf mass area and wood density}
\usage{
fuel_mass_area(LMA = 140, WD = 591, SVR = 1/0.002, partW = 0.51, partL = 0.49)
}
\arguments{
\item{LMA}{numeric. Leaf mass area in g/m2}

\item{WD}{numeric. Wood density in kg/m3}

\item{SVR}{numeric. Surface volume ratio (SVR: m2/m3), see notes.
For twigs of radius r, SVR = 2.pi.r.l*(1/2)/pi.r2.l = 1/r.
Default is for twigs of radius 2mm.}

\item{partW}{numeric. Wood mass area fraction}

\item{partL}{numeric. Leaf mass area fraction}
}
\value{
numeric. Fuel mass area in kg/m2
}
\description{
Compute fuel mass area from leaf mass area and wood density
}
