% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{get_traj}
\alias{get_traj}
\title{Get trajectory from lidar data}
\usage{
get_traj(
  las,
  thin = 1e-04,
  interval = 0.2,
  rmdup = TRUE,
  renum = FALSE,
  multi_pulse = TRUE
)
}
\arguments{
\item{las}{LAS object}

\item{thin}{LAS data thinning in seconds, see lidR::track_sensor}

\item{interval}{LAS data thinning in seconds, see lidR::track_sensor}

\item{rmdup}{If TRUE, removes duplicated points in ReturnNumber}

\item{renum}{If TRUE, renumbers points in ReturnNumber}

\item{multi_pulse}{If TRUE the pulse index is defined by (gpstime, UserData).
If FALSE the pulse index is defined by gpstime.}
}
\value{
sf object with gpstime, PointSourceID, SCORE and Point Z geometry
}
\description{
Get trajectory from lidar data
}
