% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{gpsptime_to_datetime}
\alias{gpsptime_to_datetime}
\title{convert gps time to POSIXct datetime}
\usage{
gpsptime_to_datetime(gpstime, gpstime_ref = "2011-09-14 01:46:40")
}
\arguments{
\item{gpstime}{numeric. vector of gps time}

\item{gpstime_ref}{character. reference datetime}
}
\value{
POSIXct datetime
}
\description{
convert gps time to POSIXct datetime
}
