% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{is_in_season}
\alias{is_in_season}
\title{test if points are in the months}
\usage{
is_in_season(datetime, months = 1:12)
}
\arguments{
\item{datetime}{POSIXct vector of datetime}

\item{months}{numeric vector of months defining the expected season}
}
\value{
logical vector with TRUE for points within the season
}
\description{
test if points are in the months
}
\examples{
\dontrun{
library(lidR)
filter_poi(las, is_in_season(gpstime_to_datetime(gpstime), months = 5:9))
# same as
filter_poi(las, filter_gpstime(months = 5:9))
}
}
