% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{is_near_date_mode}
\alias{is_near_date_mode}
\title{test if points are in the date mode +- deviation_days}
\usage{
is_near_date_mode(datetime, deviation_days = Inf)
}
\arguments{
\item{datetime}{POSIXct vector of datetime}

\item{deviation_days}{numeric. the number of days defining the valid range around the date mode}
}
\value{
logical vector with TRUE for points within the date mode +- deviation_days
}
\description{
test if points are in the date mode +- deviation_days
}
\examples{
\dontrun{
library(lidR)
filter_poi(las, is_near_date_mode(gpstime_to_datetime(gpstime), deviation_days = 14))
# same as
filter_poi(las, filter_gpstime(deviation_days = 14))
}
}
