% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{lasrenumber}
\alias{lasrenumber}
\title{Renumber returns by height}
\usage{
lasrenumber(las, multi_pulse = FALSE)
}
\arguments{
\item{las}{LAS object}

\item{multi_pulse}{If TRUE the pulse index is defined by (gpstime, UserData).
If FALSE the pulse index is defined by gpstime.}
}
\value{
LAS
}
\description{
Renumber returns by height
}
