% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{lasrmdup}
\alias{lasrmdup}
\title{Remove duplicated pulses or returns}
\usage{
lasrmdup(las, multi_pulse = FALSE)
}
\arguments{
\item{las}{LAS object}

\item{multi_pulse}{If TRUE the pulse index is defined by (gpstime, UserData).
If FALSE the pulse index is defined by gpstime.}
}
\value{
LAS
}
\description{
Removes in this order:
\enumerate{
\item duplicated pulses that have the different number of returns
\item pulses with more returns than NumberOfReturns
\item pulses with duplicated returns
}
}
\details{
If using both lasrdup and lasrenumber, lasrmdup should used before lasrenumber.
}
