% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_metrics.R
\name{pad_metrics}
\alias{pad_metrics}
\alias{.pad_metrics}
\title{Compute PAD metrics}
\usage{
pad_metrics(
  z0 = 0,
  dz = 1,
  nlayers = 60,
  ground_margin = 0.1,
  G = 0.5,
  omega = 0.77,
  scanning_angle = TRUE,
  cover_type = "all",
  height_cover = 2,
  use_cover = TRUE,
  limit_N_points = 0,
  limit_flight_agl = 800,
  keep_N = FALSE
)

.pad_metrics(
  gpstime,
  X,
  Y,
  Z,
  Zref,
  ReturnNumber,
  Classification,
  Easting,
  Northing,
  Elevation,
  z0 = 0,
  dz = 1,
  nlayers = 60,
  ground_margin = 0.1,
  G = 0.5,
  omega = 0.77,
  scanning_angle = TRUE,
  cover_type = "all",
  height_cover = 2,
  use_cover = TRUE,
  limit_N_points = 0,
  limit_flight_agl = 800,
  keep_N = FALSE
)
}
\arguments{
\item{z0}{numeric. Default = 0. Minimum height of the first layer in meters.}

\item{dz}{numeric. Default = 1. Height of a layer in meters.}

\item{nlayers}{numeric. Default = 60. Number of layers.}

\item{ground_margin}{numeric. Default = 0.1. Margin between the ground (Z=0) and the minimum height
of the first layer in meters. This margin is used to remove the points near the ground from the PAD
computation.}

\item{G}{numeric. Default = 0.5. Leaf projection ratio.}

\item{omega}{numeric. clumping factor.
Default is 1.
Value 1 means "no clumping" and therefore assumes a homogeneous distribution of vegetation element in the strata.
Value < 1 means clumping.}

\item{scanning_angle}{logical. Default = TRUE. Use the scanning angle computed from the trajectories to estimate cos(theta). If false: cos(theta) = 1}

\item{cover_type}{character. Default = "all". Should all returns or only first returns be considered for cover estimation.
Accepted values are be "all" or "first".}

\item{height_cover}{numeric. Default = 2. The height from which the canopy cover should be estimated.}

\item{use_cover}{logical. Default = FALSE. Use cover for PAD estimates.}

\item{limit_N_points}{numeric. Default = 400. minimum number of point in the pixel/plot for computing profiles & metrics.}

\item{limit_flight_agl}{numeric. Default = 800. Limit flight height above ground in m. If the distance between the flight height and the ground
and (Elevation - Zref) is lower than \code{limit_flight_agl}, NULL is returned.
This limit serves as a safeguard to eliminate cases where the trajectory reconstruction would be outlier.}

\item{keep_N}{logical. Default = FALSE. Keep the number of entering rays (N) and the number of hits (Ni) in each layer.}

\item{gpstime, X, Y, Z, Zref, ReturnNumber, Classification, Easting, Northing, Elevation}{numeric vectors. Attributes of the lidar point cloud.}
}
\value{
\itemize{
\item \code{pad_metrics}: a formula for lidR::xxx_metrics functions
\item \code{.pad_metrics}: a list of PAD metrics, see below.
}

The list of PAD metrics includes:
\itemize{
\item date as the mean gpstime
\item Cover layers at 2m, 4m and 6m
\item Cover_h_pad as the cover metric used for the PAD computation above \code{height_cover}.
\item cos_theta which is the average scan zenith angle used in the computation of PAD.
If \code{scanning_angle = FALSE}, \code{cos_theta} is set to 1, i.e. pulses are considered vertical.
\item PAD layers from \code{z0} to \code{z0 + nlayers * dz}, named as \verb{PAD_\{dz\}_\{zi\}} with zi the bottom edge of layer i.
}

If keep_N = TRUE, the list also contains Ni and N layers at same heights than PAD layers.
}
\description{
This function computes PAD metrics from a lidar point cloud preprocessed output.
The expected LAS attributes are: \code{gpstime}, \code{X}, \code{Y}, \code{Z}, \code{Zref}, \code{ReturnNumber}, \code{Easting},
\code{Northing}, \code{Elevation}.
}
\examples{
\dontrun{
las_file <- system.file("extdata", "M30_FontBlanche.laz", package = "lidarforfuel")
las <- lidR::readLAS(las_file)
# In real life, traj should be computed separately with buffer to avoid border effects
traj <- get_traj(las)
nlas <- fPCpretreatment(las, traj = traj)
pad <- lidR::cloud_metrics(nlas, pad_metrics(z0 = 0, dz = 0.5, nlayers = 120))
# or
pad_rast <- lidR::pixel_metrics(nlas, pad_metrics(), res = 10)
}
}
