% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_metrics.R
\name{parse_pad_heights}
\alias{parse_pad_heights}
\title{Parse dz and z_bottom from PAD names}
\usage{
parse_pad_heights(pad_names)
}
\arguments{
\item{pad_names}{A vector of PAD names of type \verb{PAD_\{dz\}_\{z_bottom\}}.}
}
\value{
A data frame with columns \code{index}, \code{dz} and \code{z_bottom},
where \code{index} is the index of the PAD layer in the pad_names vector,
\code{dz} is the layer thickness in meters, and \code{z_bottom} is the
bottom height of the layer in meters.
}
\description{
Parse dz and z_bottom from PAD names
}
\examples{
\dontrun{
pad <- lidR::cloud_metrics(nlas, pad_metrics())
parse_pad_heights(names(pad))
}
}
