% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{pixel_filter}
\alias{pixel_filter}
\title{Filter by pixel}
\usage{
pixel_filter(las, res, start = c(0, 0), filter)
}
\arguments{
\item{las}{A \code{\link[lidR]{LAS}} object.}

\item{res}{numeric. The resolution of the output. Can optionally be raster of type stars or SpatRaster.
In that case the raster is used as the template.}

\item{start}{vector of x and y coordinates for the reference raster. Default is (0,0) meaning that the grid aligns on (0,0). Not considered if res is a raster.}

\item{filter}{formula for logical predicate. Multiple conditions are combined with '&' or '|'.}
}
\value{
A \code{\link[lidR]{LAS}} object with the filtered data.
}
\description{
Filter by pixel
}
\examples{
\dontrun{
library(lidR)
library(lidarforfuel)
grid <- expand.grid(x = 1:100, y = 1:100)
las <- lidR::LAS(
  data = data.table(
    X = as.numeric(grid$x),
    Y = as.numeric(grid$y),
    Z = 0,
    # distribute points on a year
    gpstime = as.numeric(seq_len(nrow(grid))) * 3600 * 24 * 365 / nrow(grid)
  )
)
las1 <- pixel_filter(las = las, res = 10, filter = ~ filter_gpstime(gpstime, months = 5:8))
range(gpsptime_to_datetime(las$gpstime))
range(gpsptime_to_datetime(las1$gpstime))
}
}
