% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot profile & metrics}
\usage{
plot_profile(
  PADval = NULL,
  CBDval = NULL,
  zval,
  Fuel_metrics,
  plotname,
  FMAcan = 0.25,
  FMAshrub = 0.25,
  CanopyHeightThreshold = 1,
  dz = 1,
  smooth_pad = FALSE,
  bdmax = 1
)
}
\arguments{
\item{see}{ffuelmetrics2()}
}
\description{
Plot the bulk density profile and the corresponding fuel metrics
}
