% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCBDprofile_fuelmetrics.R
\name{fCBDprofile_fuelmetrics}
\alias{fCBDprofile_fuelmetrics}
\title{Fuel metrics LiDAR}
\usage{
fCBDprofile_fuelmetrics(
  X,
  Y,
  Z,
  Zref,
  Easting,
  Northing,
  Elevation,
  LMA,
  gpstime,
  threshold = 0.012,
  scanning_angle = TRUE,
  WD,
  limit_N_points = 400,
  limit_flyheight,
  datatype = "Pixel",
  omega = 1,
  d = 0.5,
  G = 0.5
)
}
\arguments{
\item{X, Y, Z, Zref}{numeric, coordinates of a point cloud (Z being the normalized Z coordinate and Zref the original one)}

\item{Easting, Northing, Elevation}{numeric, coordinates of the plane associated to each point}

\item{LMA}{numeric. Leaf mass area in g.cm² associated to each point or a generic value}

\item{threshold}{numeric or character. Default = 0.012. Bulk density critical threshold  used to discriminate strata, get CBH.etc. Either numeric : a bulk density value (in kg/m3) or character: a percentage of maximum CBD value (e.g "5%")}

\item{scanning_angle}{logical. Default = TRUE. Use the scanning angle computed from the trajectories to estimate cos(theta). If false: cos(theta) = 1}

\item{WD}{numeric. wood density associated to each point or a generic value}

\item{limit_N_points}{numeric. Default = 400. minimum number of point in the pixel/plot for computing profiles & metrics.}

\item{datatype}{character. Default is "Pixel". Either "Pixel" (if using pixel metric function) or "Plot" if only a plot is computed. Only the output change (see return). The function will be modified so it can take directly a las file when a plot only is used.}

\item{omega}{numeric. clumping factor. Default is 1. 1 mean no clumping assuming a homogeneous distribution of vegetation element}

\item{d}{numeric. default = 1. depth of the strata in meter to compute the profile}

\item{G}{numeric. Default = 0.5. Leaf projection ratio. x}

\item{norm_ground}{logical (default is FALSE). Calculate ground normals.}

\item{limit_flightheight}{numeric. Default is 400 minimum flight height above canopy in m. If the flight height is lower than limit_flyheight bulk density profile is not computed.  This limit serves as a safeguard to eliminate cases where the trajectory reconstruction would be outlier.}
}
\value{
If datatype = "Pixel" a vector containing all the fuel metrics and the CBD value for each strata
 If datatype="Plot" a list of 2 elements: 1) a vector with all fuel metrics 2) a data.table with the PAD and CBD profile value (three columns: H, PAD and CBD),
}
\description{
Function to compute PAD and CBD profiles from ALS point cloud and obtain fuel metrics from it. T
}
\details{
his function can be used with pixel_metrics lidR function to generate maps (raster). This explains why several elements of a las object have to be given as separated argument (impossible to use a las in pixel_metrics). In a following version it will be possible to use a .las file as input instead of each attribute separately
}
\examples{
to come
}
