% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCBDprofile_fuelmetrics.R
\name{fCBDprofile_fuelmetrics}
\alias{fCBDprofile_fuelmetrics}
\title{Fuel metrics LiDAR}
\usage{
fCBDprofile_fuelmetrics(
  datatype = "Pixel",
  X,
  Y,
  Z,
  Zref,
  ReturnNumber,
  Easting,
  Northing,
  Elevation,
  LMA,
  gpstime,
  height_cover = 2,
  threshold = 0.02,
  scanning_angle = TRUE,
  use_cover = FALSE,
  WD,
  limit_N_points = 400,
  limit_flight_height = 800,
  limit_vegetation_height = 0.1,
  H_PAI = 0,
  omega = 0.77,
  d = 1,
  G = 0.5
)
}
\arguments{
\item{datatype}{either "Pixel" or directly a laz/las file. Default is "Pixel". "Pixel" if the function is used with pixel_metric function to map fuel metrics. Or a .laz/.las file if a plot point cloud only needs to be computed. In the latter case, all arguments corresponding to a .laz file attributes are automatically retrieved, therefore, no need to fulfill argument X,Y,Z,Zref, Easting, Northing, Elevation, LMA, gpstime.}

\item{X, Y, Z, Zref}{numeric, coordinates of a point cloud (Z being the normalized Z coordinate and Zref the original one)}

\item{ReturnNumber}{numeric, the return number of a pulse. Used to for calculating cover.}

\item{Easting, Northing, Elevation}{numeric, coordinates of the plane associated to each point}

\item{LMA}{numeric. Leaf mass area in g/m² associated to each point or a generic value}

\item{gpstime}{gpstime of the point cloud. Only used to retrieve date of scanning}

\item{height_cover}{numeric. Default = 2. The height from which the canopy cover should be estimated.}

\item{threshold}{numeric or character. Default = 0.02. Bulk density critical threshold  used to discriminate the different strata limits, midstrorey height, canopy base, canopy top etc. Either numeric : a bulk density value (in kg/m3) or character: a percentage of maximum CBD value.}

\item{scanning_angle}{logical. Default = TRUE. Use the scanning angle computed from the trajectories to estimate cos(theta). If false: cos(theta) = 1}

\item{use_cover}{logical. Default = FALSE. Use cover for PAD estimates}

\item{WD}{numeric. wood density associated to each point or a generic value}

\item{limit_N_points}{numeric. Default = 400. minimum number of point in the pixel/plot for computing profiles & metrics.}

\item{limit_flight_height}{numeric. Default = 800. flight height above canopy in m. If the flight height is lower than limit_flyheight bulk density profile is not computed.  This limit serves as a safeguard to eliminate cases where the trajectory reconstruction would be outlier.}

\item{limit_vegetation_height}{numeric. Default = 0.1. Vegetation height below which bulkdensity profile should not be computed. NULL -> -1 is returned.}

\item{H_PAI}{Numeric. Height from which PAI, VCI and entropy is estimated. Default is 0 (means from the ground to the top).}

\item{omega}{numeric. clumping factor. Default is 1. One means "no clumping" and therefore assumes a homogeneous distribution of vegetation element in the strata.}

\item{d}{numeric. Default = 1. depth of the strata in meter to compute the profile}

\item{G}{numeric. Default = 0.5. Leaf projection ratio.}
}
\value{
If datatype = "Pixel" raster is returned with 173 Bands corresponding to metrics and bulk density profile value per strata of depth d. If datatype is a las a list of two elements: 1) a vector with all fuel metrics 2) a data.table with the PAD and CBD profile value (three columns: H, PAD and CBD),
}
\description{
Function to compute PAD and CBD profiles from a pretreated ALS point cloud (cf fPCpretreatment) and obtain fuel metrics from it.
}
\details{
This function can be used with pixel_metrics lidR function to generate maps (raster). Most of the argument of the function (i.e X,Y,Z,Zref,Easting, Northing,Elevation,LMA,WD,gpstime) comes from a pretreated poincloud obtained with the function fPCpretreatment. Note that not only fuel metrics are quantified but also: Height, plant area index above one meter (PAI_tot), vertical complexity index (VCI) based on plant area density profile or based on point cloud (lidR method). Note that the bulk density values of the profile are given in the raster using one layer per strata (with a depth = d) starting from layer 23 (i.e Band 23). Note also that in case of using the plot approach (i.e datatype = las) the profile  is given as a data.table in the second element of the list.
}
\examples{
\donttest{
path2laz <- system.file("extdata", "M30_FontBlanche_pretreated.laz", package = "lidarforfuel")

# read a pretreated las
M30_FontBlanche_pretreated <- lidR::readLAS(path2laz)

# Run the function on a las file and get a vector of the metrics and PAD and CDB profile
Fuel_metrics <- fCBDprofile_fuelmetrics(datatype = M30_FontBlanche_pretreated, WD = 500)

# Run with pixel_metrics and get a raster
M30_FontBlanche_Raster <- lidR::pixel_metrics(
  M30_FontBlanche_pretreated,
  ~ fCBDprofile_fuelmetrics(
    X = X, Y = Y, Z = Z, Zref = Zref,
    gpstime = gpstime, ReturnNumber = ReturnNumber,
    Easting = Easting, Northing = Northing, Elevation = Elevation,
    LMA = LMA, threshold = 0.016, WD = 500, limit_N_points = 400,
    datatype = "Pixel", omega = 0.77, d = 0.5, G = 0.5
  ),
  res = 10
)

# Replace -1 in cells not computed  by NA
M30_FontBlanche_Raster <- terra::subst(M30_FontBlanche_Raster, -1, NA)

# Plot a few metrics
terra::plot(M30_FontBlanche_Raster[[8:23]])
}
}
