% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffuelmetrics2.R
\name{ffuelmetrics2}
\alias{ffuelmetrics2}
\title{Fuel metrics from PAD profiles or bulk density profiles}
\usage{
ffuelmetrics2(
  PADval = NULL,
  CBDval = NULL,
  FMAcan = 0.25,
  FMAshrub = 0.25,
  smooth_pad = FALSE,
  zval,
  dz,
  LadderFuelBDthresholds = c(0.01, 0.02, 0.05),
  CanopyHeightThreshold = 1,
  CanopyBDTreshFrac = 0.3,
  CanopyTopBDThresholds = c(0.0012, 0.01),
  CBHThresholdRatio = 1/3,
  bdmax = 1,
  M = 2.5,
  as_tibble = FALSE
)
}
\arguments{
\item{PADval}{numeric vector or SpatRaster. PAD values in m2/m3. Set to NULL to input directly CBDval.}

\item{CBDval}{numeric vector or SpatRaster. CBD values in kg/m3, only used if PADval is NULL.}

\item{FMAcan}{numeric or SpatRaster. Fuel mass area in kg/m2 for canopy}

\item{FMAshrub}{numeric or SpatRaster. Fuel mass area in kg/m2 for shrub}

\item{smooth_pad}{boolean. If true PADval is smoothed by a moving average with a window of 3.
Recommended for dz=0.5 or if point number is limited}

\item{zval}{numeric vector. Bottom height of the PAD layers in m, e.g. for PAD_1_10, zval should be 10.5.}

\item{dz}{numeric. Vertical resolution of the PADval or CBDval in m, e.g. for PAD_1_10, dz should be 1.}

\item{LadderFuelBDthresholds}{numeric vector. Minimal bulk density thresholds for Ladder fuels (kg/m3)}

\item{CanopyHeightThreshold}{numeric. Default minimal value of height in m for canopy/ladder (versus surface fuel) to look for CBH and CTH, typically 1 to 3 m (default 1m)}

\item{CanopyBDTreshFrac}{numeric. Fraction threshold to combine CBDmin and CBDmax to look for CBH (0.5 is average, 0.1 is close to CBDmin)}

\item{CanopyTopBDThresholds}{numeric vector. Minimal bulkDensity to look for CTH (canopy height but in terms of fuel<Height)}

\item{CBHThresholdRatio}{numeric. Ratio of total height above which we look for CBH}

\item{bdmax}{numeric. Max value of bulk density kg/m3 for canopy (above CanopyHeightThreshold=1 m only)}

\item{M}{numeric. M factor from Perrakis 2025}

\item{as_tibble}{logical. If TRUE, returns a tibble. Not used in the raster case.}
}
\value{
a data frame, a tibble or a SpatRaster.
}
\description{
Fuel metrics from PAD profiles or bulk density profiles
}
