% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fPCpretreatment.R
\name{fPCpretreatment}
\alias{fPCpretreatment}
\title{Point cloud pre-treatment for using fCBDprofile_fuelmetrics in pixels}
\usage{
fPCpretreatment(chunk, classify = F, norm_ground = F, LMA)
}
\arguments{
\item{chunk}{path a las (laz) file. Can be apply to a catalog see lidr catalog apply)}

\item{classify}{logical (default is FALSE). Make a ground classification. Only if the original point cloud is not classified}

\item{norm_ground}{logical (default is FALSE). Calculate ground normals.}

\item{LMA}{Path to a LMA map (.tif) of the area if available or a single LMA value in g.m² (e.g 140. cf: Martin-Ducup et al. 2024)}
}
\value{
a Normalyzed point cloud (.laz) with several new attributes need to run fCBDprofile_fuelmetrics
}
\description{
Function to do the pretreatment on las (laz) files for using in fCBDprofile_fuelmetrics (get CBD profiles & fuel metrics). This can be used in catalog_apply lidR function. The pre-treatment consist in adding to noramlyze the pointcloud and adding several attribute: plane position for each point (easting, northin, elevation) and LMA by crossing the pointcloud with a LMA map
}
\details{
The attributes added to the laz are LMA : LMA value of each point. Zref :original Z; Easting, Northing, Elevation, Time that are the X,Y,Z position of the plane and the its GPStime for each point (obtained from lidR::track_sensor()). In a following version it will be possible to directly load a trajectory file if available.
}
\examples{
\donttest{
path2laz=system.file("extdata","M30_FontBlanche.laz", package="lidarforfuel")
 #LMA value selected = 120.6 that is the LMA for Pinus halepensis, the dominant species of the plot
M30_FontBlanche_pretreated<-fPCpretreatment(path2laz,LMA=120.6)
# displaying the new attributes in the las
names(M30_FontBlanche_pretreated)
}
}
