% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad_metrics}
\alias{pad_metrics}
\title{Compute PAD metrics}
\usage{
pad_metrics(
  z0 = 0,
  dz = 1,
  nlayers = 60,
  G = 0.5,
  omega = 0.77,
  scanning_angle = TRUE,
  height_cover = 2,
  cover_type = "all",
  limit_N_points = 400,
  limit_flight_height = 800,
  keep_N = FALSE
)
}
\arguments{
\item{z0}{numeric. Default = 0. Minimum height of the first layer in meters.}

\item{dz}{numeric. Default = 1. Height of a layer in meters.}

\item{nlayers}{numeric. Default = 60. Number of layers.}

\item{G}{numeric. Default = 0.5. Leaf projection ratio.}

\item{omega}{numeric. clumping factor.
Default is 1.
Value 1 means "no clumping" and therefore assumes a homogeneous distribution of vegetation element in the strata.
Value < 1 means clumping.}

\item{scanning_angle}{logical. Default = TRUE. Use the scanning angle computed from the trajectories to estimate cos(theta). If false: cos(theta) = 1}

\item{height_cover}{numeric. Default = 2. The height from which the canopy cover should be estimated.}

\item{cover_type}{character. Default = "all". Should all, first or no returns be considered for cover estimation.
Accepted values are be "all", "first" or NULL. If NULL, cover estimation is not used in PAD computation.}

\item{limit_N_points}{numeric. Default = 400. minimum number of point in the pixel/plot for computing profiles & metrics.}

\item{limit_flight_height}{numeric. Default = 800. Limit flight height above ground in m. If the flight height is lower than
limit_flight_height, NULL is returned.
This limit serves as a safeguard to eliminate cases where the trajectory reconstruction would be outlier.}

\item{keep_N}{logical. Default = FALSE. Keep the number of entering rays (N) and the number of hits (Ni) in each layer.}
}
\value{
A list of PAD metrics
PAD layers, 5 Cover layers and mean gpstime
If keep_N = TRUE, the list also contains Ni and N layers.
}
\description{
This function computes PAD metrics from a lidar point cloud preprocessed output.
}
\examples{
\donttest{
 las_file <-   system.file("extdata", "M30_FontBlanche.laz", package = "lidarforfuel")
 las <- lidR::readLAS(las_file)
 # In real life, traj should be done computed with buffer to avoid border effects
 traj <- get_traj(las)
 nlas <- fPCpretreatment(las, traj = traj)
 pad <- lidR::cloud_metrics(nlas, pad_metrics(z0 = 0, dz = 0.5, nlayers = 120))
 # or
 pad_rast <- lidR::pixel_metrics(nlas, pad_metrics(), res = 10)
}
}
